+++ parser/src/test/ui_tests.rs
@@ -1,5 +1,6 @@
 ui_test!(error_empty_enum);
 ui_test!(grapheme_width);
+ui_test!(invalid_event_id);
 ui_test!(invalid_function_id);
 ui_test!(recursive1);
 ui_test!(recursive2_a);