+++ parser/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Raise an error on invalid function ids.
+- Raise an error on invalid event ids.
 
 ## [0.4.0] - 2024-03-21